Update 3.2-linear-regression-models-and-least-squares.ipynb
Since X is a N x (p+1) matrix (with the additional columns of 1s for the intercept \beta_0), the trace of H is trace(X(X^TX)^{-1}X^T) = trace(I_{p+1}) = (p+1). That's how you get that \widehat{\sigma^2}, conditional on X, is an unbiased estimator of \sigma^2.

By law of expectations, this also implies \widehat{\sigma^2} is an unconditionally unbiased estimator of \sigma^2.